reorganised file names and reorganised user settings
